Move testhelpers into their own package - cleans up some duplication in the tests - cleans up release builds imports of the utils package
